zstd release 0.8.0 - made an extra zstd.mak file - removed Format7zZStd - SFXSetup also has zstd now - the big 7z.dll is backward compatible to: 0.5 .. 0.7 - the small 7za.dll has only support for zstd 0.8

+++ CPP/7zip/Compress/ZstdDecoder.h
@@ -1,8 +1,6 @@
 // ZstdDecoder.h
 // (C) 2016 Rich Geldreich, Tino Reichardt
 
-#include "StdAfx.h"
-
 #define ZSTD_STATIC_LINKING_ONLY
 #include "../../../C/Alloc.h"
 #include "../../../C/ZStd/zstd.h"
@@ -62,8 +60,8 @@ class CDecoder:public ICompressCoder,
 
   UInt32 _inBufSizeAllocated;
   UInt32 _outBufSizeAllocated;
-  UInt32 _inBufSize;
-  UInt32 _outBufSize;
+  size_t _inBufSize;
+  size_t _outBufSize;
 
   HRESULT CreateBuffers ();
   HRESULT CodeSpec (ISequentialInStream * inStream, ISequentialOutStream * outStream, ICompressProgressInfo * progress);
